+++ include/resources/Resource.h
@@ -6,9 +6,9 @@
 // Struct to define a type of resource (like Grass)
 struct ResourceType {
     std::string name;
-    int max_amount; // Maximum amount of resource on a tile
-    int regrowth_rate; // Amount that regrows per turn
-    int nutritional_value; // Nutritional value of the resource
+    float max_amount; // Maximum amount of resource on a tile
+    float regrowth_rate; // Amount that regrows per turn
+    float nutritional_value; // Nutritional value of the resource
     char symbol; // Symbol to draw on the grid
 };
